Refactor: SocketProvider(context) 폴더구조 변경 & 불필요한 코드 삭제
