+++ src/main.tsx
@@ -5,7 +5,7 @@ import theme from './styles/theme';
 import GlobalStyle from './styles/GlobalStyles';
 import { RecoilRoot } from 'recoil';
 import './styles/fonts/font.css';
-import { SocketProvider } from './recoil/SocketProvider.tsx';
+import { SocketProvider } from './context/SocketProvider.tsx';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
 	<ThemeProvider theme={theme}>
