--- src/recoil/SocketState.ts
@@ -1,7 +1,0 @@
-import { atom } from 'recoil';
-import { Socket } from 'socket.io-client';
-
-export const SocketStateAtom = atom<Socket | null>({
-	key: 'SocketStateAtom',
-	default: null,
-});
